RWSemapクラスを別ファイルに分離
